refactor segmentation app flow and fix test suite
